requirements.txt
gspread==3.1.0
oauth2client==4.1.3
riotwatcher==2.5.0
tensorflow-gpu==1.13.1
Keras==2.2.4
